--- a/benchmarks/src/main/java/reactor/core/publisher/FluxBufferTimeoutBenchmark.java
+++ b/benchmarks/src/main/java/reactor/core/publisher/FluxBufferTimeoutBenchmark.java
@@ -97,8 +97,4 @@ public class FluxBufferTimeoutBenchmark {
 			countDown();
 		}
 	}
-
-	public static void main(String[] args) {
-
-	}
 }