The fair backpressure variant of the `bufferTimeout` operator has been reworked to use a state machine with a minimum number of volatile variables eliminating potential data races, such as skipping the delivery when `onNext` and timeout happen concurrently or cancellation happens while `onNext` is delivered, etc.

Resolves #3531