@OlegDokuka 2023-11-08
Although the above could be false negative, since the async task can be unawaited properly. Can you please doublecheck @chemicL